Move LinearUtil to DimensionReduction from DimensionReduction/Delaunay
