+++ src/main/scala/DimensionReduction/package.scala
@@ -5,7 +5,7 @@
  *  The subpackages of this package are [[DBSCAN]] and [[Delaunay]]. The former package provides an implementation of
  *  the DBSCAN algorithm, while the latter package provides algorithms that produce the Delaunay simplicialization and
  *  convex hull of a set of points. The [[Delaunay]] package also has a Linear Algebra helper object,
- *  [[Delaunay.LinearUtil]]
+ *  [[LinearUtil]]
  *
  */
 package object DimensionReduction
